requirements.txt
django
django-tastypie
django-crispy-forms
requests
